Small improvements, and better modal management on auto-popup

+++ src/main/services/cloud-service.js
@@ -61,6 +61,7 @@ class CloudService {
   }
 
   async disconnect () {
+    if (!this.socket) return
     await this.socket.destroy({ force: true })
   }
 }
